Remove duplicate definition of PopularityAsc
Which resulted in warnings displayed whenever the library is required.

The reason it is appearing twice is that because it is defined twice in
ebaySvc.xsd (schema) file and class generator doesn't check uniqueness
of enumerable values.

Perhaps eBay going to remove it from future schema, since it says:
> This field is deprecated (in first definition)
> This value is not used (in second definition)

The following command can be used to quickly find those definitions:
> ack 'PopularityAsc' -C 6 lib/ebay/schema/ebaySvc.xsd